ingestion/requirements.txt
python-dotenv
pymilvus
openai 
tiktoken 
langchain
unstructured